Refactor the grid building
Previously, the grid code was all jammed together - web fetching,
parsing and widget building were all happening in the same class.

Create a new LetterGrid class to do the widget building. This class
streams the users updates back to the caller. The streaming requires a
StatefulWidget so that the StreamController can be cleaned up.