Add highlight of the currently selected clue.
Not totally sure about this commit as:
a) I'm not sure the structure is correct.
b) For some reason it looks a bit naff.

This commit:

 - Parses the span for each clue sent by the server.
 - Constructs a map from cell Index to all the other cell Indexes that
are part of the same clue (only for across clues so far).
 - adds an "onFocus" callback to the Cells. When they are focused,
this callback looks up which cells should now be highlighted, and this is
set as State on the LetterGrid. The children are then rebuilt with the correct
highlighting.

It seems likely the onFocus behaviour will be needed for moving the focus
around as well.