Make entire screen a SingleChildScrollView
This commit is an admission of horrible failure...

I cannot for the life of me get the independent scrolling of both the
clues and the grid working without overflow (grid scrolling only for
times when the bottom is covered by the keyboard).

Revert to behaviour where the user can scroll the whole screen. This is
slightly less bad UX than it used to be because now the active clues are
displayed at the top of the screen. Once the "tap clue to select in
grid" is implemented it will be even better.