Refactor: Move clue state up to crossword_model.dart
Previously, the current clue state was stored in letter_grid.dart. This
meant:
1) The active clue header had to be built as part of the grid
2) We couldn't access the state for future features (like scrolling to active
clue)

Move the state of the current clue up to StaticCrossword.
Move the building of the clue header up as well.
Delete the redundat Cursor data type, it becomes Index everywhere.

N.B. This breaks cursor auto-increment. Not 100% sure why but I think
it's because the grid (and therefore Cell children) is now being rebuilt?
Breaking is fine though, because it didn't really work before, and I'm
going to fix it anyway.