Fix multiple cursors issue.
Give each Cell a FocusNode as part of its State. Request focus on that
FocusNode depending on if the cell is focused. Delete any dealing with
FocusNode from LetterGrid.

This massively simplifies the code, and makes more sense. In fact, you
might say that this is the obviously correct way to do it and eveything
before was nonsense.