Make entries appear immediately.
Make TextEditingController part of the Cell State. Then when the user
updates the value, mark the cell NOT to take the server value when it is
rebuilt (which happens immediately to change the highlighting).

Also, fix a bug with cursor advancement - the entire app broke at the
end of words otherwise.