Implement responsive design
Choose layout based on screen width, and render [Grid, Across Clues,
Down Clues] as a row if there is space. Otherwise render them as a
scrolling column.

Should now render ok on the most common phone and desktop/laptop sizes.
Still not great if you are (e.g.) using half a monitor or a tablet,
though.